requirements.txt
torchvision
pytorch_lightning
torch
nltk
flask
numpy
requests
python-dotenv